requirements.txt
# LangChain Learning Project Requirements
# ================================================

# Core LangChain packages
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50

# Environment management
python-dotenv>=1.0.0

# Data manipulation and analysis
pandas>=1.3.0

# Jupyter notebook support (for IPython.display)
ipython>=7.0.0

# Additional LangChain integrations
langchain-hub>=0.1.0

# Google AI dependencies (if needed)
google-generativeai>=0.3.0